Added icons and hover animation
Added plus icon that would later on toggle the input form up and down into and out of visibility. Also added trash cans instead of the previous "X" characters that can be used to delete items from the to do list. Added transition to the trash cans so that they are normally hidden and only fade into view when the list item is being hovered on.